+++ ui/ui.ts
@@ -356,8 +356,12 @@ function submit_video_wrapper(...args) {
 }
 
 function submit_postprocessing(...args) {
-  log('SubmitExtras');
+  const id = randomId();
+  log('SubmitProcess', id);
   clearGallery('extras');
+  requestProgress(id, null, null);
+  window.submit_state = '';
+  args[0] = id;
   return args;
 }